@@ -1,8 +1,8 @@
 package cn.mlus.bettervannilafishes.client.animator;
 
 import cn.mlus.bettervannilafishes.entity.nautilus.NautilusEntity;
+import software.bernie.geckolib.animation.AnimationState;
 import software.bernie.geckolib.cache.object.GeoBone;
-import software.bernie.geckolib.core.animation.AnimationState;
 import software.bernie.geckolib.model.GeoModel;
 
 public class BvfNautilusAnimator extends GeneralAnimator<NautilusEntity>{
