@@ -1,6 +1,5 @@
 package cn.mlus.bettervannilafishes.data;
 
-import cn.aurorian.ers.init.ErsTagKeys;
 import cn.mlus.bettervannilafishes.BetterVannilaFishes;
 import cn.mlus.bettervannilafishes.init.BvfItems;
 import cn.mlus.bettervannilafishes.init.BvfTagKeys;
@@ -9,7 +8,8 @@ import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.level.block.Block;
-import net.minecraftforge.common.data.ExistingFileHelper;
+
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -24,27 +24,27 @@ public class BvfItemTagProvider extends ItemTagsProvider{
     @Override
     protected void addTags(HolderLookup.@NotNull Provider provider) {
         this.tag(ItemTags.FISHES)
-                .add(BvfItems.HADDOCK_COD.get())
-                .add(BvfItems.ATLANTIC_COD.get())
-                .add(BvfItems.PACIFIC_COD.get())
-                .add(BvfItems.MALE_SALMON.get())
-                .add(BvfItems.FEMALE_SALMON.get())
-                .add(BvfItems.PACIFIC_SALMON.get())
-                .add(BvfItems.YELLOW_FIN_PUFFER.get())
-                .add(BvfItems.OBSCURE_PUFFER.get());
+                .add(BvfItems.HADDOCK_COD.value())
+                .add(BvfItems.ATLANTIC_COD.value())
+                .add(BvfItems.PACIFIC_COD.value())
+                .add(BvfItems.MALE_SALMON.value())
+                .add(BvfItems.FEMALE_SALMON.value())
+                .add(BvfItems.PACIFIC_SALMON.value())
+                .add(BvfItems.YELLOW_FIN_PUFFER.value())
+                .add(BvfItems.OBSCURE_PUFFER.value());
 
-        this.tag(ErsTagKeys.KNOWN_FISH)
-                .add(BvfItems.HADDOCK_COD.get())
-                .add(BvfItems.ATLANTIC_COD.get())
-                .add(BvfItems.PACIFIC_COD.get())
-                .add(BvfItems.MALE_SALMON.get())
-                .add(BvfItems.FEMALE_SALMON.get())
-                .add(BvfItems.PACIFIC_SALMON.get())
-                .add(BvfItems.YELLOW_FIN_PUFFER.get())
-                .add(BvfItems.OBSCURE_PUFFER.get());
+//        this.tag(ErsTagKeys.KNOWN_FISH)
+//                .add(BvfItems.HADDOCK_COD.value())
+//                .add(BvfItems.ATLANTIC_COD.value())
+//                .add(BvfItems.PACIFIC_COD.value())
+//                .add(BvfItems.MALE_SALMON.value())
+//                .add(BvfItems.FEMALE_SALMON.value())
+//                .add(BvfItems.PACIFIC_SALMON.value())
+//                .add(BvfItems.YELLOW_FIN_PUFFER.value())
+//                .add(BvfItems.OBSCURE_PUFFER.value());
 
         this.tag(BvfTagKeys.PUFFER)
-                .add(BvfItems.YELLOW_FIN_PUFFER.get())
-                .add(BvfItems.OBSCURE_PUFFER.get());
+                .add(BvfItems.YELLOW_FIN_PUFFER.value())
+                .add(BvfItems.OBSCURE_PUFFER.value());
     }
 }
