@@ -12,14 +12,15 @@ public class NautilusBlessingEffect extends MobEffect {
     }
 
     @Override
-    public void applyEffectTick(@NotNull LivingEntity pLivingEntity, int pAmplifier) {
+    public boolean applyEffectTick(@NotNull LivingEntity pLivingEntity, int pAmplifier) {
         pLivingEntity.setAirSupply(pLivingEntity.getMaxAirSupply());
         if(pLivingEntity.hasEffect(MobEffects.DIG_SLOWDOWN))
             pLivingEntity.removeEffect(MobEffects.DIG_SLOWDOWN);
+        return true;
     }
 
     @Override
-    public boolean isDurationEffectTick(int pDuration, int pAmplifier) {
+    public boolean shouldApplyEffectTickThisTick(int p_295329_, int p_295167_) {
         return true;
     }
 }
