@@ -2,7 +2,7 @@ package cn.mlus.bettervannilafishes.entity;
 
 import cn.mlus.bettervannilafishes.client.animator.GeneralAnimator;
 import net.minecraft.world.entity.Mob;
-import software.bernie.geckolib.core.animatable.GeoAnimatable;
+import software.bernie.geckolib.animatable.GeoAnimatable;
 
 public interface BvfEntity<T extends Mob & GeoAnimatable> {
     GeneralAnimator<T> getAnimator();
