@@ -12,6 +12,6 @@ public class PacificCod extends BvfCodEntity {
         super(pEntityType, pLevel);
     }
     public @NotNull ItemStack getBucketItemStack() {
-        return new ItemStack(BvfItems.PACIFIC_COD_BUCKET.get());
+        return new ItemStack(BvfItems.PACIFIC_COD_BUCKET);
     }
 }
