@@ -4,16 +4,16 @@ import net.minecraft.core.particles.DustParticleOptions;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
+import net.minecraft.tags.EntityTypeTags;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.MobType;
 import net.minecraft.world.entity.monster.Guardian;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BowItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.event.ForgeEventFactory;
+import net.neoforged.neoforge.event.EventHooks;
 import org.jetbrains.annotations.NotNull;
 import org.joml.Vector3f;
 
@@ -21,7 +21,7 @@ import javax.annotation.Nullable;
 
 public class EnchantedPuffer extends BowItem {
     public EnchantedPuffer(Properties pProperties) {
-        super(pProperties.defaultDurability(50));
+        super(pProperties.durability(50));
     }
 
     private int clientSideAttackTime;
@@ -44,7 +44,7 @@ public class EnchantedPuffer extends BowItem {
         LivingEntity closestTarget = null;
         double closestDistance = Double.MAX_VALUE;
         for (LivingEntity entity : pPlayer.level().getEntitiesOfClass(LivingEntity.class, pPlayer.getBoundingBox().inflate(range))) {
-            if (entity != pPlayer && (entity.getMobType().equals(MobType.UNDEAD) || entity instanceof Guardian)) {
+            if (entity != pPlayer && (entity.getType().is(EntityTypeTags.UNDEAD) || entity instanceof Guardian)) {
                 double distance = pPlayer.distanceToSqr(entity);
                 if (distance < closestDistance) {
                     closestDistance = distance;
@@ -119,8 +119,8 @@ public class EnchantedPuffer extends BowItem {
 
     public void releaseUsing(@NotNull ItemStack pStack, @NotNull Level pLevel, @NotNull LivingEntity pEntityLiving, int pTimeLeft) {
         if (pEntityLiving instanceof Player player) {
-            int i = this.getUseDuration(pStack) - pTimeLeft;
-            i = ForgeEventFactory.onArrowLoose(pStack, pLevel, player, i, true);
+            int i = this.getUseDuration(pStack,pEntityLiving) - pTimeLeft;
+            i =  EventHooks.onArrowLoose(pStack, pLevel, player, i, true);
             if (i < 0) {
                 return;
             }
@@ -128,7 +128,7 @@ public class EnchantedPuffer extends BowItem {
             float f = getPowerForTime(i);
             if (!((double)f < 0.1)) {
                 if (!pLevel.isClientSide) {
-                    pStack.hurtAndBreak(1, player, (p_289501_) -> p_289501_.broadcastBreakEvent(player.getUsedItemHand()));
+                    pStack.hurtAndBreak(1, player, LivingEntity.getSlotForHand(pEntityLiving.getUsedItemHand()));
                     if(clientSideCachedAttackTarget != null)
                         clientSideCachedAttackTarget.hurt(pLevel.damageSources().magic(), 20.0F * f);
                 }
